# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ConsultationInput {
  /** @minLength 1 */
  message: string;
}

export interface ConsultationAnswerInput {
  /** @minLength 1 */
  answer: string;
}

export type ConsultationRiskLevel = typeof ConsultationRiskLevel[keyof typeof ConsultationRiskLevel];


export const ConsultationRiskLevel = {
  green: 'green',
  yellow: 'yellow',
  orange: 'orange',
  red: 'red',
} as const;

export type ConsultationStatus = typeof ConsultationStatus[keyof typeof ConsultationStatus];


export const ConsultationStatus = {
  in_progress: 'in_progress',
  complete: 'complete',
  emergency: 'emergency',
} as const;

export interface Consultation {
  id: string;
  createdAt: string;
  updatedAt: string;
  title: string;
  symptomPreview: string;
  riskLevel: ConsultationRiskLevel;
  riskLabel: string;
  status: ConsultationStatus;
  progress: number;
  /** @nullable */
  nextQuestion: string | null;
}

export type ConsultationMessageRole = typeof ConsultationMessageRole[keyof typeof ConsultationMessageRole];


export const ConsultationMessageRole = {
  assistant: 'assistant',
  user: 'user',
  system: 'system',
} as const;

export interface ConsultationMessage {
  id: string;
  role: ConsultationMessageRole;
  content: string;
  timestamp: string;
}

export type CaseSummaryRiskLevel = typeof CaseSummaryRiskLevel[keyof typeof CaseSummaryRiskLevel];


export const CaseSummaryRiskLevel = {
  green: 'green',
  yellow: 'yellow',
  orange: 'orange',
  red: 'red',
} as const;

export interface CaseSummary {
  chiefComplaint: string;
  symptoms: string[];
  duration: string;
  severity: string;
  associatedSymptoms: string[];
  medicalHistory: string;
  possibleConditions: string[];
  confidenceScore: number;
  riskLevel: CaseSummaryRiskLevel;
  riskReason: string;
  recommendedAction: string;
  selfCareAdvice: string[];
  doctorRecommendation: string;
  disclaimer: string;
}

export type ConsultationDetail = Consultation & ({
  messages: ConsultationMessage[];
  summary: CaseSummary | null;
});

export type DashboardSummaryRiskBreakdown = {
  green: number;
  yellow: number;
  orange: number;
  red: number;
};

export interface DashboardSummary {
  totalConsultations: number;
  completedConsultations: number;
  latestConsultation: Consultation | null;
  averageRisk: string;
  recentReports: Consultation[];
  riskBreakdown: DashboardSummaryRiskBreakdown;
}

export interface WorkflowAgent {
  id: string;
  name: string;
  role: string;
  description: string;
  input: string;
  output: string;
  order: number;
}

export interface Report {
  consultation: Consultation;
  summary: CaseSummary;
  generatedAt: string;
}