# lib/api-zod/src/generated/types/caseSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CaseSummaryRiskLevel } from './caseSummaryRiskLevel';

export interface CaseSummary {
  chiefComplaint: string;
  symptoms: string[];
  duration: string;
  severity: string;
  associatedSymptoms: string[];
  medicalHistory: string;
  possibleConditions: string[];
  confidenceScore: number;
  riskLevel: CaseSummaryRiskLevel;
  riskReason: string;
  recommendedAction: string;
  selfCareAdvice: string[];
  doctorRecommendation: string;
  disclaimer: string;
}