# lib/api-zod/src/generated/types/consultation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConsultationRiskLevel } from './consultationRiskLevel';
import type { ConsultationStatus } from './consultationStatus';

export interface Consultation {
  id: string;
  createdAt: string;
  updatedAt: string;
  title: string;
  symptomPreview: string;
  riskLevel: ConsultationRiskLevel;
  riskLabel: string;
  status: ConsultationStatus;
  progress: number;
  /** @nullable */
  nextQuestion: string | null;
}