# lib/api-zod/src/generated/types/consultationAnswerInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConsultationAnswerInput {
  /** @minLength 1 */
  answer: string;
}