# lib/api-zod/src/generated/types/consultationDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CaseSummary } from './caseSummary';
import type { Consultation } from './consultation';
import type { ConsultationMessage } from './consultationMessage';

export type ConsultationDetail = Consultation & ({
  messages: ConsultationMessage[];
  summary: CaseSummary | null;
});