# lib/api-zod/src/generated/types/consultationMessage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConsultationMessageRole } from './consultationMessageRole';

export interface ConsultationMessage {
  id: string;
  role: ConsultationMessageRole;
  content: string;
  timestamp: string;
}