# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Consultation } from './consultation';
import type { DashboardSummaryRiskBreakdown } from './dashboardSummaryRiskBreakdown';

export interface DashboardSummary {
  totalConsultations: number;
  completedConsultations: number;
  latestConsultation: Consultation | null;
  averageRisk: string;
  recentReports: Consultation[];
  riskBreakdown: DashboardSummaryRiskBreakdown;
}