# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CaseSummary } from './caseSummary';
import type { Consultation } from './consultation';

export interface Report {
  consultation: Consultation;
  summary: CaseSummary;
  generatedAt: string;
}