# lib/api-zod/src/generated/types/workflowAgent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkflowAgent {
  id: string;
  name: string;
  role: string;
  description: string;
  input: string;
  output: string;
  order: number;
}